+++ bin/main.ml
@@ -27,6 +27,7 @@ let all : _ Cmdliner.Cmd.t list =
     ; Ocaml_merlin.command
     ; Shutdown.command
     ; Diagnostics.command
+    ; Targets_cmd.command
     ]
   in
   let groups =